EDU-736 Extract logic of storage-location core
