+++ src/components/files-viewer.js
@@ -5,11 +5,7 @@ import classNames from 'classnames';
 import FilesGridViewer from './files-grid-viewer.js';
 import FilesListViewer from './files-list-viewer.js';
 import { cdnObjectShape } from '../ui/default-prop-types.js';
-
-export const FILES_VIEWER_DISPLAY = {
-  grid: 'grid',
-  list: 'list'
-};
+import { FILES_VIEWER_DISPLAY } from '../domain/constants.js';
 
 function FilesViewer({
   display,
