+++ src/components/resource-picker/resource-selector.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
 import { useStorage } from '../storage-context.js';
 import React, { useEffect, useState } from 'react';
-import StorageLocation from './storage-location.js';
+import StorageLocation from './storage-location-screens-manager.js';
 import { STORAGE_LOCATION_TYPE } from '../../domain/constants.js';
 import { getStorageLocationTypeForUrl } from '../../utils/storage-utils.js';
