+++ src/components/resource-picker/resource-selector.less
@@ -1,3 +1,4 @@
 .ResourceSelector {
+  height: 70vh;
   position: relative;
 }
