+++ src/components/resource-picker/storage-location.less
@@ -23,11 +23,11 @@
   justify-content: space-between;
 }
 
-.StorageLocation-filesViewerSelectContainer {
+.StorageLocation-selectContainer {
   min-width: 100px;
 }
 
-.StorageLocation-filesViewerSelect {
+.StorageLocation-select {
   width: 100%;
 }
